flask_app/requirements_1.txt
tensorflow-cpu==2.4.1
Keras==2.4.3
numpy==1.19.2
pandas==1.1.5
opencv-python-headless==4.5.1.48
Pillow==8.1.2